[Refactor] test search movie success
